Introduce more generic methods for casting
Mostly see cast.rs.

After a bunch of research and discussion (in
https://users.rust-lang.org/t/why-does-vec-from-raw-parts-require-same-size-and-not-same-size-capacity/73036)
I now believe that these are safe to do (relatively safe; see inline
docs).

This also includes marking a lot of structs with `Copy` to indicate that
they don’t implement `Drop`.

This mostly obviates the need for a custom allocator like we talked
about in https://github.com/Zaplib/zaplib/issues/103. We only need that
for casts like `cast_vec::<u32, u8>(..)`, where the two types have
different alignments. But that’s less common. We were also thinking of
using that for https://github.com/Zaplib/zaplib/issues/105, by always
casting to `u8`, and then back to the desired type. But we can just
implement that ticket in a more mechanical way; it’s probably not worth
having a custom allocator just for that!